feat: Migrates to next-mdx-enhanced
