+++ .gitignore
@@ -71,3 +71,6 @@ yarn-error.log
 
 # Vercel
 .vercel
+
+# Ignore next-mdx-enhanced cache directory
+.mdx-data
