+++ next.config.js
@@ -1,14 +1,33 @@
 const withSourceMaps = require("@zeit/next-source-maps")();
+const readingTime = require("reading-time");
+const mdxPrism = require("mdx-prism");
+const withMdxEnhanced = require("next-mdx-enhanced");
 
-module.exports = withSourceMaps({
-  webpack: (config, { isServer }) => {
-    // Fixes npm packages that depend on `fs` module
-    if (!isServer) {
-      config.node = {
-        fs: "empty",
-      };
-    }
+module.exports = withSourceMaps(
+  withMdxEnhanced({
+    layoutPath: "./components/layouts/articles",
+    defaultLayout: true,
+    remarkPlugins: [
+      require("remark-autolink-headings"),
+      require("remark-slug"),
+      require("remark-code-titles"),
+    ],
+    rehypePlugins: [mdxPrism],
+    extendFrontMatter: {
+      process: (mdxContent) => ({
+        wordCount: mdxContent.split(/\s+/gu).length,
+        readingTime: readingTime(mdxContent),
+      }),
+    },
+  })({
+    webpack: (config, { isServer }) => {
+      if (!isServer) {
+        config.node = {
+          fs: "empty",
+        };
+      }
 
-    return config;
-  },
-});
+      return config;
+    },
+  })
+);
