+++ package.json
@@ -2,7 +2,7 @@
   "name": "nirmalyaghosh.com",
   "description": "Personal portfolio of Nirmalya Ghosh",
   "author": "Nirmalya Ghosh <nirmalya.email@gmail.com>",
-  "version": "0.0.4",
+  "version": "0.0.5",
   "private": false,
   "license": "MIT",
   "keywords": [
@@ -23,21 +23,23 @@
     "@chakra-ui/core": "^0.8.0",
     "@emotion/core": "^10.0.28",
     "@emotion/styled": "^10.0.27",
+    "babel-plugin-import-glob-array": "^0.2.0",
     "emotion-theming": "^10.0.27",
     "gray-matter": "^4.0.2",
+    "mdx-prism": "^0.3.1",
     "next": "9.4.4",
+    "next-mdx-enhanced": "^3.0.0",
     "next-seo": "^4.5.0",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-icons": "^3.10.0",
     "react-lazyload": "^2.6.7",
+    "reading-time": "^1.2.0",
     "recoil": "^0.0.7",
-    "rehype-highlight": "^4.0.0",
-    "rehype-stringify": "^7.0.0",
-    "remark-parse": "^8.0.2",
-    "remark-react": "^7.0.1",
-    "remark-rehype": "^7.0.0",
-    "unified": "^9.0.0"
+    "rehype": "^10.0.0",
+    "remark-autolink-headings": "^6.0.0",
+    "remark-code-titles": "^0.1.1",
+    "remark-slug": "^6.0.0"
   },
   "devDependencies": {
     "@sentry/browser": "^5.16.1",
