+++ pages/articles.tsx
@@ -1,0 +1,17 @@
+import Page from "components/pages/articles/base";
+import { NextPage } from "next";
+import Head from "next/head";
+import { frontMatter as articles } from "./articles/*.mdx";
+
+const ArticlesIndexPage: NextPage = () => {
+  return (
+    <>
+      <Head>
+        <title>Articles</title>
+      </Head>
+      <Page articles={articles} />
+    </>
+  );
+};
+
+export default ArticlesIndexPage;
