+++ types/article.ts
@@ -1,8 +1,7 @@
 export default interface Article {
   id: number;
   title: string;
-  banner: string;
   date: string;
   tags: string[];
-  processedContent: string;
+  __resourcePath: string;
 }
